Fixed Version Handler obsolete files dialog on refresh.
In some cases an asset database refresh does not result in
a reload of the app domain (e.g non-DLL files are changed)
in which case the obsolete files dialog was not being displayed.
This changes the Version Handler to cache whether obsolete files
were found and an update should be performed when the asset
database refresh is complete, at which point the obsolete files
dialog is displayed prompting users to clean up old plugin
components.

Change-Id: I82796786615dbeb64d462dd5591c52c0b01e702f