Updated Version Handler manifests to most recent installed files.
Improved global tracking of obsolete files across VH package
upgrades. Previously if a file was deemed obsolete due to the
difference in one package's manifest it would not have been
detected as obsolete in a dependent package. This commit changes
manifest parsing to search for all revisions of each file and
select the latest revision in all parsed manifests.

Bug: 150471207
Change-Id: I07587b86fca18ec60741f9eb6f10fd2b0c06f374