Improved VH manifest alias tracking for packages without aliases.
I5c459bb092666088c3eedb33155bf06df75b6285 added an empty set of
aliases to all manifests including those that are actually aliases
this had the side effect of breaking the alias graph traversal.
This patch only adds an empty set of aliases to manifests who are
not aliases of other manifests.

Change-Id: I83a195c7f17864f0ce01e7e7a99d717ddeb631ce