Fixed export_unity_package.py on macOS.
Python's tarfile and gzip on macOS generates a .tar.gz that can't be
read by Unity on Windows. This changes export_unity_package.py to
use BSD tar distributed with macOS to generate tarballs instead
which not only can be expanded on Windows but also speeds up the
archiving process.

Fixed #351

Change-Id: Ifc82666c6496f9a11088b6a7ea01e9cbad593846