Make sure MultiSelectWindow for different purposes are unique.
EditorWindow.GetWindow() only create a new window or return an
existing one per type.  As a result, when CreateMultiSelectWindow()
is called for different purposes at the same time, the same window
instance can be filled with data from the different code path, and
causes error.

This change make sure the MultiSelectWindow created for different
purposes are unique.

Change-Id: I7dafdb49ef714def221afcc359561ec5cff886ed