control the events queue when -executeMethod is provided
When -executeMethod is provided, some versions of Unity might not
trigger an editor update event. This causes our callbacks queue (to
detect changes in build type) to keep growing at a rate faster than
consumption/execution of those callbacks. This causes Unity to freeze
on opening a project with "-executeMethod" provided on the cmd line.

Bug: 162245206
Fixes #389

Change-Id: Ib6924ab778dca915a12f9b682091b5c1a36db2ee