Fixed generation of .unitypackage with tarfile.
When using gzip and tarfile python modules to create a
.unitypackage, by default the gzip module will place the
output filename in the gzip header which results in Unity
on Windows (just Windows!) refusing to list the contents of
the file. To workaround this issue, this change replaces
the output filename in the gzip header with OUTPUT.tar.gz
when writing, for example, OUTPUT.unitypackage.

Tested:
* All unit test pass.
* Verified this works with Unity 2018.4.17f1 on Windows and
  still works in Unity on macOS and Linux.

Bug: 165661723
Change-Id: I38580b98e44a23c8eeb71790f57184901f7d46fd