attempt to resolve before a build
Due to changes in newer versions of Unity, plugin modules are not reloaded
before build and this meant that some of our optimizations to resolve
only when necessary didn't work as expected. With this fix, we will be
running a resolve every build (provided there were no build errors in prior
attempts), and also run it only on one scene for a build with multiple scenes.
Jetifier enabling via gradle properties wasn't flagged as a build setting that
is checked for differences during "are there differences for a new resolve?".

Bug: 170395178
Fixes: #810
Change-Id: I2b8994497bf48c2eb6276137289927859e10f64f