Add a new param to GetXcodeTargetGuids()
Add a new param to GetXcodeTargetGuids() to return guids from all
targets in Unity 2019.3+.

By default, GetXcodeTargetGuids() only returns guids of the target which
includes Unity and third-party libraries. It is `Unity-iPhone` for Unity
2019.2 or below or `UnityFramework` for Unity 2019.3+.

Introduce a new param "includeAllTargets" to return guids of all
targets, excluding test targets, when multiple Xcode project target is
supported from Unity 2019.3+.

This is useful when we need to add target membership to all targets for
special cases.  For example, adding `Google-Services.plist` to all
targets in order to support the case when the user's project has
dependency on Swift framework, such as Facebook SDK.

Bug: 174515725

Change-Id: I2374800cefc21d79cf9a7fb6f66a39f0acbbdade