Changed verioned dll to a version folder (Fix #441)
Versioned dll files such as IOSResolver.dll are named with version
postfix such as `_v1.2.166` in the file name.
This is causing issues with Unity 2021.1.11+ to load some libraries due
to `Assembly name not match`.

In order to workaround with Unity's issue, this change modified the
build script to put those versioned dlls to be under a folder named
after the version number, ex.
  `ExternalDependencyManager/Editor/1.2.166/Google.IOSResolver.dll`