Remove version folder when generating meta file template
During release (`./gradlew release`), the build script copy the generated
`.meta` files back to `plugin/Asset` folder to be used as the templates for
future builds, before next release.

This change make sure the version folder is removed from the path in the staging
folder before copying back to the template folder.