Remove the Always embed swift libraries flag
As part of enabled swift, we had forced ALWAYS_EMBED_SWIFT_STANDARD_LIBRARIES to YES, but that should not be necessary, as we also add an empty swift file to the target, which will enable swift. On top of that, the always embed flag seems to cause problems for some when submitting apps because of the additional framework folder it might make.