Add log when Unity failed to create folder
Follow up for #543. When Unity failed to create folder using Unity API
`AssetDatabase.CreateFolder()`, even with the workaround in #543, Unity
would still generate two error message and there is no way to turn it
off.

This change add an additional log after Unity API fails to inform the
user to ignore this error.