Improve XmlDependencies.IsDependenciesFile performance
This simpler string manipulation checks the same logic as before but much faster in the context of a large project. If the OnPostprocessAllAssets callback is invoked multiple times with thousands of files, as can happen with non-trivial Unity projects, the regex call becomes quite expensive.

For profiling data see: https://github.com/googlesamples/unity-jar-resolver/issues/601